code finish Note and CashFlow

+++ YouTrade.Winform/YouTrade.Winform/MainForm.cs
@@ -23,7 +23,7 @@ namespace YouTrade.Winform
         string pathIn = "", pathOut = "";
         string pathTempIncome = "", pathTempBasicInfo = "";
         string pathTempRatios = "", pathTempBalance = "", pathTempStock = "";
-        string pathTempCashFlow = "";
+        string pathTempCashFlow = "", pathTempNote = "";
         DataSet dsSource = null;
         DataSet dsSource1 = null;
         int demIncome = 0, demBasicInfo = 0;
@@ -96,6 +96,12 @@ namespace YouTrade.Winform
             if(!Directory.Exists(pathTempCashFlow))
             {
                 Directory.CreateDirectory(pathTempCashFlow);
+            }
+            //Path Temp Note
+            pathTempNote = System.Windows.Forms.Application.StartupPath + "\\Output\\Note\\Temp\\";
+            if (!Directory.Exists(pathTempNote))
+            {
+                Directory.CreateDirectory(pathTempNote);
             }
         }
         #endregion
@@ -379,6 +385,51 @@ namespace YouTrade.Winform
                 }
             }
         }
+        //Note
+        void MoveToTempNote()
+        {
+            // Chuyển file sang folder Temp
+            var files = Directory.GetFiles(tbInput.Text, "*.*", SearchOption.AllDirectories).Where(s => s.EndsWith(".xls") || s.EndsWith(".xlsm") || s.EndsWith(".xlsx")).Where(f => f.Contains("note") && !f.Contains("~$"));
+            Microsoft.Office.Interop.Excel.Application excelApp = null;
+            Microsoft.Office.Interop.Excel.Workbook excelWorkbook = null;
+            foreach (string file in files)
+            {
+                try
+                {
+                    excelApp = new Microsoft.Office.Interop.Excel.Application();
+                    excelApp.FileValidation = MsoFileValidationMode.msoFileValidationSkip;
+                    string fileName = Path.GetFileNameWithoutExtension(file);
+                    string fileEx = Path.GetExtension(file);
+                    string FullNameIn = tbInput.Text + fileName + fileEx;
+                    string fullNameIn_In_Temp = pathTempNote + fileName.Replace(".", string.Empty) + ".xls";
+                    if (!File.Exists(fullNameIn_In_Temp))
+                    {
+                        excelWorkbook = excelApp.Workbooks.Open(FullNameIn, 1, false, 5, "", "", false, Microsoft.Office.Interop.Excel.XlPlatform.xlWindows, "", true, false, null, false);
+                        excelApp.DisplayAlerts = false;
+                        string fileNameOut = pathTempNote + fileName.Replace(".", string.Empty);
+                        excelWorkbook.SaveAs(fileNameOut, Microsoft.Office.Interop.Excel.XlFileFormat.xlWorkbookNormal, Type.Missing, Type.Missing, false, false, Microsoft.Office.Interop.Excel.XlSaveAsAccessMode.xlNoChange, Type.Missing, Type.Missing, Type.Missing, Type.Missing, Type.Missing);
+                    }
+                }
+                catch (Exception ex)
+                {
+                }
+                finally
+                {
+                    if (excelWorkbook != null)
+                    {
+                        Marshal.FinalReleaseComObject(excelWorkbook);
+                        excelWorkbook = null;
+                    }
+                    if (excelApp != null)
+                    {
+                        excelApp.Quit();
+                        Marshal.FinalReleaseComObject(excelApp);
+                        excelApp = null;
+                    }
+                }
+            }
+        }
+
         #endregion
         #region Save to DB
         private void SaveToDBRatio(System.Data.DataTable dt)
@@ -619,6 +670,65 @@ namespace YouTrade.Winform
             }
 
         }
+        //Note
+        private void SaveToDBNote(System.Data.DataTable dt)
+        {
+            using (SqlConnection dbcon = new SqlConnection(sqlConnectionString))
+            {
+                dbcon.Open();
+                for (int i = 8; i <= dt.Rows.Count - 1; i++)
+                {
+                    try
+                    {
+                        if (dt.Rows[i][1].ToString().Trim() == "")
+                            break;
+                        for (int j = 4; j <= dt.Columns.Count - 1; j++)
+                        {
+                            string strq = "INSERT INTO [dbo].[Note] ([Ticker],[Year] ,[Quater]  ,[Name] ,[Value] ,[Unit]) VALUES(@ticker,@year,@quarter,@feildname,@value,@unit)";
+                            SqlCommand sqlcmdD = new SqlCommand(strq, dbcon);
+
+                            sqlcmdD.Parameters.AddWithValue("@ticker", dt.Rows[i][1].ToString().Trim());
+                            string pattern = dt.Rows[6][j].ToString();
+                            // Get Name
+                            int startPositionName = pattern.IndexOf(".") + ".".Length;
+                            string patternT = pattern.Substring(startPositionName, pattern.Length - startPositionName);
+
+                            int startPositionNameDot = patternT.IndexOf(".") + ".".Length;
+                            string patternTDot = patternT.Substring(startPositionNameDot, patternT.Length - startPositionNameDot);
+
+                            string name = patternTDot.Substring(0, patternTDot.IndexOf("Consolidated\nYear:"));
+
+                            // Get Year
+                            int startPositionYear = patternTDot.IndexOf("Consolidated\nYear:") + "Consolidated\nYear:".Length;
+                            string year = patternTDot.Substring(startPositionYear, patternTDot.IndexOf("Quarter") - startPositionYear);
+
+                            // Get Quarter
+                            int startPositionQuarter = patternTDot.IndexOf("Quarter") + "Quarter:".Length;
+                            string quarter = patternTDot.Substring(startPositionQuarter, patternTDot.IndexOf("Unit") - startPositionQuarter);
+
+                            //Get unit
+                            int startPositionUnit = patternTDot.IndexOf("Unit") + "Unit:".Length;
+                            string unit = patternTDot.Substring(startPositionUnit, patternTDot.Length - startPositionUnit);
+                            //Annual
+
+                            sqlcmdD.Parameters.AddWithValue("@feildname", name.ToString().Trim());
+                            sqlcmdD.Parameters.AddWithValue("@year", Convert.ToInt16(year));
+                            sqlcmdD.Parameters.AddWithValue("@quarter", quarter.ToString().Trim() != "Annual" ? Convert.ToInt16(quarter.ToString().Trim()) : 5);
+                            sqlcmdD.Parameters.AddWithValue("@value", dt.Rows[i][j].ToString().Trim());
+                            sqlcmdD.Parameters.AddWithValue("@unit", unit.ToString().Trim());
+                            sqlcmdD.ExecuteNonQuery();
+                            if (dt.Rows[i][1].ToString().Trim() == "")
+                                break;
+                        }
+                    }
+                    catch
+                    {
+                    }
+                }
+                dbcon.Close();
+            }
+
+        }
         #endregion
 
         #region Store file
@@ -647,6 +757,20 @@ namespace YouTrade.Winform
                     Directory.CreateDirectory(tbOutput.Text + "\\cashflow\\");
                 }
                 File.Copy(fileName, tbOutput.Text + "\\cashflow\\" + filenameOnly, true);
+                File.Delete(fileName);
+            }
+        }
+        //Note
+        private void StoreFileNote(string fileName)
+        {
+            if (File.Exists(fileName))
+            {
+                string filenameOnly = Path.GetFileName(fileName);
+                if (!Directory.Exists(tbOutput.Text + "\\Note\\"))
+                {
+                    Directory.CreateDirectory(tbOutput.Text + "\\Note\\");
+                }
+                File.Copy(fileName, tbOutput.Text + "\\Note\\" + filenameOnly, true);
                 File.Delete(fileName);
             }
         }
@@ -715,6 +839,19 @@ namespace YouTrade.Winform
         private void tbInput_TextChanged(object sender, EventArgs e)
         {
 
+        }
+
+        private void MainForm_Load(object sender, EventArgs e)
+        {
+
+        }
+
+        private void btnNote_Click(object sender, EventArgs e)
+        {
+            btnNote.Text = "Note Running...";
+            MoveToTempNote();
+           ReadExcelAndSaveNote();
+            btnNote.Text = "Note";
         }
         #endregion
 
@@ -868,7 +1005,7 @@ namespace YouTrade.Winform
                     {
                         string fileName = Path.GetFileNameWithoutExtension(file);
 
-                        string fullNameIn_In_Out = tbOutput.Text + "cashflow\\" + fileName.Replace(".", string.Empty) + ".xls";
+                        string fullNameIn_In_Out = tbOutput.Text + "Cashflow\\" + fileName.Replace(".", string.Empty) + ".xls";
                         if (!File.Exists(fullNameIn_In_Out))
                         {
 
@@ -881,6 +1018,40 @@ namespace YouTrade.Winform
 
                         }
                         StoreFileCashFlow(file);
+                    }
+                    catch
+                    {
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+            }
+        }
+        void ReadExcelAndSaveNote()
+        {
+            try
+            {
+                var files1 = Directory.GetFiles(pathTempNote, "*.*", SearchOption.AllDirectories).Where(s => s.EndsWith(".xls"));
+                foreach (string file in files1)
+                {
+                    try
+                    {
+                        string fileName = Path.GetFileNameWithoutExtension(file);
+
+                        string fullNameIn_In_Out = tbOutput.Text + "note\\" + fileName.Replace(".", string.Empty) + ".xls";
+                        if (!File.Exists(fullNameIn_In_Out))
+                        {
+
+                            dsSource = GetDatasetFromExcel(file);
+                            foreach (System.Data.DataTable tbl in dsSource.Tables)
+                            {
+                                SaveToDBNote(tbl);
+                                break;
+                            }
+
+                        }
+                        StoreFileNote(file);
                     }
                     catch
                     {
